requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Vector Search
faiss-cpu>=1.7.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.0

# LLM Integration
openai>=1.50.0
ollama>=0.3.0

# Web Framework & API
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# UI Framework
streamlit>=1.30.0
streamlit-chat>=0.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
tiktoken>=0.5.0
tqdm>=4.65.0

# Scientific Computing
numpy>=1.24.0
scikit-learn>=1.3.0

# Other essentials
PyYAML>=6.0
packaging>=23.0
typing_extensions>=4.5.0